requirements-api.txt
fastapi
uvicorn
mlflow
pandas
python-dotenv